some modif

+++ ERPWebAPI.EL/Concrete/OHS/OHS_MachineryRecordType.cs
@@ -8,7 +8,7 @@ namespace ERPWebAPI.EL.Concrete.OHS
     public class OHS_MachineryRecordType :IEntity
     {
         [Key]
-        public short RecordTypeId { get; set; }
+        public byte RecordTypeId { get; set; }
         public string RecordType { get; set; }
         public string? UserEmployee { get; set; }
         public DateTime? TransactionDate { get; set; }
